+++ src/MapboxPlugin.php
@@ -41,7 +41,7 @@ class MapboxPlugin extends Plugin
     /**
      * @var string Current schema version of the plugin.
      */
-    public string $schemaVersion = '1.0.0';
+    public string $schemaVersion = '1.1.0';
 
     /**
      * @var MapboxPlugin Self-referential plugin property.